+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - [x] Rank groups by WoT
 - [x] Fix subtle navigation bugs
 - [x] Show a little more metadata on relay card
+- [x] Make groups deletable
 
 # 0.4.4